+++ src/context/PhysicsContext.ts
@@ -8,6 +8,7 @@ type PhysicsContextProps = {
   register: (body: Matter.Body, fn: PhysicsUpdateFunction) => () => void;
   setGravity: (velocity: Velocity) => void;
   setVelocity: (body: Matter.Body, velocity: Velocity) => void;
+  applyForce: (body: Matter.Body, position: Position, velocity: Velocity) => void;
   addEventListener: (type: PhysicsEventType, fn: (event: PhysicsEvent) => void) => void;
   removeEventListener: (type: PhysicsEventType, fn: (event: PhysicsEvent) => void) => void;
 }
@@ -17,6 +18,7 @@ export const PhysicsContext = React.createContext<PhysicsContextProps>({
   register: () => () => {},
   setGravity: () => {},
   setVelocity: () => {},
+  applyForce: () => {},
   addEventListener: () => {},
   removeEventListener: () => {},
 });